refactor: organize legacy algorithm imports
