+++ anomavision/padim/__init__.py
@@ -1,6 +1,6 @@
 """Backward-compatible imports for the modular PaDiM implementation."""
 
-from .algorithm.padim.padim import *
-from .algorithm.padim.padim import Padim
+from ..algorithm.padim.padim import *
+from ..algorithm.padim.padim import Padim
 
 __all__ = ["Padim"]
