+++ anomavision/patchcore/__init__.py
@@ -1,6 +1,6 @@
 """Backward-compatible imports for the modular PatchCore implementation."""
 
-from .algorithm.patchcore.patchcore import *
-from .algorithm.patchcore.patchcore import PatchCore, build_patchcore_from_stats
+from ..algorithm.patchcore.patchcore import *
+from ..algorithm.patchcore.patchcore import PatchCore, build_patchcore_from_stats
 
 __all__ = ["PatchCore", "build_patchcore_from_stats"]
